+++ features/transaction/atc_common/src/main/java/com/tokopedia/atc_common/domain/model/response/AddToCartMultiDataModel.kt
@@ -31,7 +31,7 @@ data class AtcMultiData (
                     val cartId: Long = -1,
 
                     @SerializedName("product_id")
-                    val productId: Int = -1,
+                    val productId: Long = -1,
 
                     @SerializedName("quantity")
                     val quantity: Int = -1,
