+++ features/transaction/oneclickcheckout/src/main/java/com/tokopedia/oneclickcheckout/order/view/processor/OrderSummaryPagePromoProcessor.kt
@@ -135,7 +135,7 @@ class OrderSummaryPagePromoProcessor @Inject constructor(private val validateUse
         val ordersItem = Order()
         ordersItem.shopId = orderCart.shop.shopId.toLong()
         ordersItem.uniqueId = orderCart.cartString
-        ordersItem.product_details = listOf(ProductDetail(orderCart.product.productId.toLong(), orderCart.product.quantity.orderQuantity))
+        ordersItem.product_details = listOf(ProductDetail(orderCart.product.productId, orderCart.product.quantity.orderQuantity))
         ordersItem.isChecked = true
 
         ordersItem.shippingId = shipping.getRealShipperId()
